Refactor(backend): Move the schema item path helpers out of SearchService
