+++ backend/Origam.Architect.Server/Services/SearchService.cs
@@ -86,7 +86,7 @@ public class SearchService(
     {
         try
         {
-            ISchemaItem root = GetRoot(item);
+            ISchemaItem root = SchemaItemTreePath.GetRoot(item);
             return new SearchResult
             {
                 SchemaId = item.Id,
@@ -96,7 +96,7 @@ public class SearchService(
                 Folder = root.Group?.Path ?? "",
                 Package = item.PackageName,
                 PackageReference = referencePackages.Contains(item.SchemaExtensionId),
-                ParentNodeIds = GetParentNodeIds(item, root),
+                ParentNodeIds = SchemaItemTreePath.GetParentNodeIds(item, root),
                 IsOrphaned = false,
             };
         }
@@ -116,67 +116,6 @@ public class SearchService(
                 FoundIn = item.Name ?? item.Id.ToString(),
                 IsOrphaned = true,
             };
-        }
-    }
-
-    public static ISchemaItem GetRoot(ISchemaItem item)
-    {
-        try
-        {
-            ISchemaItem root = item;
-            for (ISchemaItem parent = item.ParentItem; parent != null; parent = parent.ParentItem)
-            {
-                root = parent;
-            }
-            return root;
-        }
-        catch (Exception ex)
-        {
-            throw new OrphanedSchemaReferenceException(item.Id, ex);
-        }
-    }
-
-    public static List<string> GetParentNodeIds(ISchemaItem item, ISchemaItem root)
-    {
-        try
-        {
-            if (root.RootProvider is not AbstractSchemaItemProvider provider)
-            {
-                return [];
-            }
-
-            var ids = new List<string>();
-            AddFolderNameIfAny(ids, item);
-
-            for (ISchemaItem parent = item.ParentItem; parent != null; parent = parent.ParentItem)
-            {
-                ids.Add(parent.Id.ToString());
-                AddFolderNameIfAny(ids, parent);
-            }
-
-            for (SchemaItemGroup group = root.Group; group != null; group = group.ParentGroup)
-            {
-                ids.Add(group.Id.ToString());
-            }
-
-            ids.Add(provider.NodeId);
-            ids.Add(provider.Group);
-            ids.Reverse();
-
-            return ids;
-        }
-        catch (Exception ex)
-        {
-            throw new OrphanedSchemaReferenceException(item.Id, ex);
-        }
-
-        static void AddFolderNameIfAny(List<string> target, ISchemaItem schemaItem)
-        {
-            var folderName = schemaItem?.GetType().SchemaItemDescription()?.FolderName;
-            if (!string.IsNullOrWhiteSpace(folderName))
-            {
-                target.Add(folderName);
-            }
         }
     }
 }
